Migrate the database from Neon Postgres to MySQL on Hostinger
Neon's monthly data-transfer quota was exhausted and took the site down;
every query, including pg_dump, returned XX000. Hostinger offers no managed
Postgres, so the app moves to the MySQL/MariaDB already included in the
hosting plan.

Schema (26 tables): uuid -> varchar(36) generated in JS via newId(), since
MySQL has no server-side default — raw INSERTs now pass id explicitly.
timestamp -> datetime(3) rather than MySQL TIMESTAMP, which is capped at 2038
and rewrites values across time zones; the pool pins timezone "Z" and raw SQL
uses UTC_TIMESTAMP, never now(). Every indexed/unique/PK text column became
varchar(191), the utf8mb4-safe index width.

Runtime: 19 .returning() call sites rewritten (MySQL has none) — uuid tables
mint the id up front, AUTO_INCREMENT tables use $returningId, and UPDATE/DELETE
read affectedRows. 12 upserts moved to onDuplicateKeyUpdate, reading the row
back by its unique key because MySQL keeps the existing id on conflict.
onConflictDoNothing -> INSERT IGNORE.

Raw SQL: db.execute now returns [rows, fields], so every query goes through
rawRows() — the previous casts type-checked while handing back field metadata.
FILTER (WHERE) -> COUNT(CASE WHEN), to_char -> date_format, || -> concat(),
translate() accent folding -> collate utf8mb4_unicode_ci (not the MySQL-8-only
0900_ai_ci, so MariaDB works), desc nulls last -> plain desc, ::casts dropped,
and `key` back-quoted as a MySQL reserved word.

Verified with scripts/verify-mysql-parity.ts: 28 assertions against both a
local MySQL 8 and the live MariaDB 11.8. Data copied with
scripts/migrate-neon-to-mysql.ts and compared row by row — all 26 tables and
10,741 rows byte-identical. Neon is untouched and preserved as
SOURCE_DATABASE_URL.

Also corrects the privacy policy, which still named Neon as the database
processor.

+++ scripts/diagnose-euroleague-nulls.ts
@@ -45,7 +45,7 @@ async function main() {
     /* ---- players (current EL season) ---- */
     const playerCols = await colsFor("players")
     const [{ total: playerTotal }] = await sql<{ total: number }[]>`
-      select count(distinct p.id)::int as total
+      select count(distinct p.id) as total
       from players p
       join player_season_stats pss on pss.player_id = p.id
       join leagues l on l.id = pss.league_id
@@ -55,7 +55,7 @@ async function main() {
     console.log(`\n=== PLAYERS (EL current season): ${playerTotal} ===`)
     for (const col of playerCols) {
       const [{ n }] = await sql<{ n: number }[]>`
-        select count(distinct p.id)::int as n
+        select count(distinct p.id) as n
         from players p
         join player_season_stats pss on pss.player_id = p.id
         join leagues l on l.id = pss.league_id
@@ -69,7 +69,7 @@ async function main() {
     /* ---- player_season_stats (current EL season) ---- */
     const statCols = await colsFor("player_season_stats")
     const [{ total: statTotal }] = await sql<{ total: number }[]>`
-      select count(*)::int as total
+      select count(*) as total
       from player_season_stats pss
       join leagues l on l.id = pss.league_id
       join seasons s on s.id = pss.season_id
@@ -78,7 +78,7 @@ async function main() {
     console.log(`\n=== PLAYER_SEASON_STATS (EL current season): ${statTotal} rows ===`)
     for (const col of statCols) {
       const [{ n }] = await sql<{ n: number }[]>`
-        select count(*)::int as n
+        select count(*) as n
         from player_season_stats pss
         join leagues l on l.id = pss.league_id
         join seasons s on s.id = pss.season_id
@@ -91,7 +91,7 @@ async function main() {
     /* ---- teams (any EL team) ---- */
     const teamCols = await colsFor("teams")
     const [{ total: teamTotal }] = await sql<{ total: number }[]>`
-      select count(distinct t.id)::int as total
+      select count(distinct t.id) as total
       from teams t
       join player_season_stats pss on pss.team_id = t.id
       join leagues l on l.id = pss.league_id
@@ -101,7 +101,7 @@ async function main() {
     console.log(`\n=== TEAMS (EL current season): ${teamTotal} ===`)
     for (const col of teamCols) {
       const [{ n }] = await sql<{ n: number }[]>`
-        select count(distinct t.id)::int as n
+        select count(distinct t.id) as n
         from teams t
         join player_season_stats pss on pss.team_id = t.id
         join leagues l on l.id = pss.league_id
@@ -115,7 +115,7 @@ async function main() {
     /* ---- coaches (EL) ---- */
     const coachCols = await colsFor("coaches")
     const [{ total: coachTotal }] = await sql<{ total: number }[]>`
-      select count(*)::int as total
+      select count(*) as total
       from coaches c
       join leagues l on l.id = c.league_id
       where l.slug = 'euroleague'
@@ -123,7 +123,7 @@ async function main() {
     console.log(`\n=== COACHES (EL): ${coachTotal} ===`)
     for (const col of coachCols) {
       const [{ n }] = await sql<{ n: number }[]>`
-        select count(*)::int as n
+        select count(*) as n
         from coaches c
         join leagues l on l.id = c.league_id
         where l.slug = 'euroleague' and ${sql(`c.${col}`)} is null
@@ -134,7 +134,7 @@ async function main() {
     /* ---- team_season_stats (EL current season) ---- */
     const tssCols = await colsFor("team_season_stats")
     const [{ total: tssTotal }] = await sql<{ total: number }[]>`
-      select count(*)::int as total
+      select count(*) as total
       from team_season_stats tss
       join leagues l on l.id = tss.league_id
       join seasons s on s.id = tss.season_id
@@ -143,7 +143,7 @@ async function main() {
     console.log(`\n=== TEAM_SEASON_STATS (EL current season): ${tssTotal} rows ===`)
     for (const col of tssCols) {
       const [{ n }] = await sql<{ n: number }[]>`
-        select count(*)::int as n
+        select count(*) as n
         from team_season_stats tss
         join leagues l on l.id = tss.league_id
         join seasons s on s.id = tss.season_id
